PPTP-1061 : Enrol current user for specific PPT reference
We use the tax-enrolments service to enrol the current user for the specific PPT reference after successful PPT subscription. Since - at this time - we have no facility to send emails, we do nothing other than write to the application log when we are notified as to the eventual success or failure of this enrolment. We return details of whether the initial request to enrol the user has been successful or not so that the FE might notify the user as appropriate. Note also that we always notify NRS of the PPT subscription whether the initial enrolment has been successful or not.