PPTP-1076 : OrganisationDetails model refactor (#107)
* PPTP-1076 : OrganisationDetails model refactor

* PPTP-1076 : Make BV status optional

* PPTP-1076 : Fixup after merging main

* PPTP-1525 : WIP : Migrate Registrations

Any existing in-flight registrations which are found to be in the older format and automatically upgraded to the newer format on application startup. Note that this is only performed when the RegistrationUpgrader singleton is created; you will need to run the application in 'production' mode for this to happen.

* PPTP-1525 : WIP : Formatting!

* PPTP-1525 : WIP : Ensure logging is comprehensive

* PPTP-1525 : WIP : Fix broken test

* PPTP-1525 : WIP : Exclude temp code from coverage reports

* PPTP-1525 : WIP : Ensure Migration On Startup

Added new play module to explicitly load the RegistrationUpgrader. The migrations stopped working on startup and may have been as a result of removing of the temporary controller which I was using previously to test this functionality

* PPTP-1525 : WIP : Use booleans for flags!

* PPTP-1525 : WIP : Renamed eager loading module

* PPTP-1076 : Reintroduce param name