PPTP-1803 : RegWithoutID Fix (#142)
Since LegalEntityDetails.regWithoutID is always true for group subscriptions, we can (when creating a Registration from a Subscription) use the Registration equivalent field (OrganisationDetails.regWithoutIDFlag) to hold the value for the representative group member (the first GroupPartnershipDetails instance). When re-creating the Subscription from the amended Registration we will set LegalEntityDetails.regWithoutID to true (since it's always this value) and use the value in OrganisationDetails.regWithoutIDFlag to populate the representative group member (the first GroupPartnershipDetails instance). Other group members can then be populated from the GroupMember instances.

Ensure that we only ever set LegalEntityDetails.regWithoutIDFlag when we are creating group registration variations.